Refactored grid components
